add: Experiment workload generator
Helper script to create workloads for the 3 experiments to be
featured in our slideshow and final paper.

TODO Rewrite as a Bash file or something; this really only needs
to be ran once and committed; the method of making the file only
needs to be committed for posterity, not tacked-on like this.